+++ PGPatcher/src/main.cpp
@@ -14,6 +14,7 @@
 #include "PGRunCache.hpp"
 #include "PGUI.hpp"
 #include "common/BethesdaGame.hpp"
+#include "patchers/PatcherMeshGlobalRecalculateBounds.hpp"
 #include "patchers/PatcherMeshPostFixSSS.hpp"
 #include "patchers/PatcherMeshPostHairFlowMap.hpp"
 #include "patchers/PatcherMeshPostRestoreDefaultShaders.hpp"
@@ -636,6 +637,10 @@ void mainRunnerPrep(const ParallaxGenCLIArgs& args,
         Logger::debug("Adding Hair Flow Map post-patcher");
         meshPatchers.postPatchers.emplace_back(PatcherMeshPostHairFlowMap::factory());
     }
+    if (params.globalPatcher.isRecalculateBoundsEnabled) {
+        Logger::debug("Adding Recalculate Bounds global patcher");
+        meshPatchers.globalPatchers.emplace_back(PatcherMeshGlobalRecalculateBounds::factory());
+    }
 
     const PatcherUtil::PatcherTextureSet texPatchers;
     PGPatcher::loadPatchers(meshPatchers, texPatchers);
@@ -753,7 +758,8 @@ void mainRunnerPatch(const ParallaxGenCLIArgs& args,
                            params.processing.allowedModelRecordTypes,
                            true,
                            args.excludeFacegens,
-                           progressCallback);
+                           progressCallback,
+                           params.globalPatcher.isRecalculateBoundsEnabled);
 
     progressWindow->CallAfter([progressWindow] {
         progressWindow->setMainLabel(pgTr("progress.steps.patchingTextures"));